feat(admin-settings): move Frontend Runtime (reCAPTCHA/Clarity) para Integrations
"Frontend Runtime" era um menu jargão e isolado em Admin Settings, escondendo o
que na real são integrações de serviços de terceiros do front (Google reCAPTCHA,
Microsoft Clarity). Move pra dentro da tela de Integrations, como seção própria.

- Nova FrontendServicesSection (self-contained: useForm + load/save de
  frontend_runtime), renderizada DEPOIS das seções OAuth, com divisória — fora do
  array INTEGRATIONS (reCAPTCHA/Clarity são site key/project id, não OAuth apps).
- Remove o navItem frontendRuntime, a rota /settings/admin/frontend-runtime e a
  page FrontendRuntimeConfig (conteúdo migrou pra seção).
- Backend intocado: frontend_runtime já está na allowlist CONFIG_TYPES; mesmas keys.
- i18n nas 6 línguas: rótulos jargão → claros (Analytics & Segurança; reCAPTCHA &
  Clarity); remove navigation.frontendRuntime.
- Testes: spec da seção (7) + spec do Integrations mocka a seção p/ isolar; 24/24.

Coordenado com a mudança de Integrations OAuth desta mesma branch (ambas mexem em IntegrationsConfig.tsx).
