+++ src/i18n/locales/pt/adminSettings.json
@@ -7,8 +7,7 @@
     "channels": "Canais",
     "openai": "OpenAI",
     "integrations": "Integracoes",
-    "inboundEmail": "E-mail de Entrada",
-    "frontendRuntime": "Frontend Runtime"
+    "inboundEmail": "E-mail de Entrada"
   },
   "email": {
     "title": "Configuracao de E-mail",
@@ -556,10 +555,10 @@
     "saving": "A guardar..."
   },
   "frontendRuntime": {
-    "title": "Configuração Frontend Runtime",
-    "description": "Configure as chaves do reCAPTCHA e Clarity que são carregadas pelo frontend em tempo de execução.",
+    "title": "Analytics e Segurança",
+    "description": "Chaves de serviços front-end de terceiros: Google reCAPTCHA (proteção anti-bot) e Microsoft Clarity (analytics).",
     "fields": {
-      "cardTitle": "Chaves Runtime",
+      "cardTitle": "reCAPTCHA e Clarity",
       "recaptchaSiteKey": "Chave do Site reCAPTCHA",
       "clarityProjectId": "ID do Projeto Clarity"
     },
