+++ src/pages/Admin/Settings/index.tsx
@@ -1,6 +1,6 @@
 import { NavLink, Outlet, Navigate, useLocation } from 'react-router-dom';
 import { useLanguage } from '@/hooks/useLanguage';
-import { Mail, MailOpen, HardDrive, KeyRound, MessageSquare, Sparkles, Puzzle, Globe, Cable } from 'lucide-react';
+import { Mail, MailOpen, HardDrive, KeyRound, MessageSquare, Sparkles, Puzzle, Cable } from 'lucide-react';
 
 const navItems = [
   { key: 'email', path: '/settings/admin/email', icon: Mail },
@@ -11,7 +11,6 @@ const navItems = [
   { key: 'integrations', path: '/settings/admin/integrations', icon: Puzzle },
   { key: 'evolutionHub', path: '/settings/admin/evolution-hub', icon: Cable },
   { key: 'inboundEmail', path: '/settings/admin/inbound-email', icon: MailOpen },
-  { key: 'frontendRuntime', path: '/settings/admin/frontend-runtime', icon: Globe },
 ] as const;
 
 export default function AdminSettingsLayout() {
